backend/pyproject.toml
[tool.poetry]
name = "testworks"
version = "0.1.0"
description = ""
authors = ["Christian Lyons <christian.lyons@pm.me>"]
license = "MIT"

packages = [
    {include = "app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.29"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
pyodbc = "^5.1.0"
uvicorn = "^0.29.0"
redis = "^5.0.4"
httpx = "^0.27.0"
psycopg2 = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
coverage = "^7.5.1"
websockets = "^12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.2"
ruff = "^0.4.2"
pytest-cov = "^5.0.0"
pytest-html = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
